wallet export cleanup
