+++ src/app/lookups/wallet-export-csv-columns.lookup.ts
@@ -1,0 +1,17 @@
+export const CsvColumns = [
+  { header: 'Transaction Hash', property: 'transactionHash' },
+  { header: 'Transaction Event No.', property: 'transactionEventNumber' },
+  { header: 'Block Number', property: 'blockNumber' },
+  { header: 'Block Time', property: 'blockTime' },
+  { header: 'Account', property: 'account' },
+  { header: 'Gas Fee (CRS)', property: 'gasFeeCrs' },
+  { header: 'Gas Fee (Fiat)', property: 'gasFeeFiat' },
+  { header: 'Transaction Type', property: 'transactionType' },
+  { header: 'Fiat Currency', property: 'currency' },
+  { header: 'Amount Spent', property: 'amountSpent' },
+  { header: 'Token Spent', property: 'tokenSpent' },
+  { header: 'Total Fiat Spent', property: 'totalFiatSpent' },
+  { header: 'Amount Received', property: 'amountReceived' },
+  { header: 'Token Received', property: 'tokenReceived' },
+  { header: 'Total Fiat Received', property: 'totalFiatReceived' },
+];
